% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp}
\usage{
mycltp(n, iter, lambda = 10)
}
\arguments{
\item{n}{Sample size}

\item{iter}{Number of iterations}

\item{lambda}{Lambda parameter for Poisson distribution, demonstrating CLT}
}
\value{
A CLT Poisson distribution
}
\description{
mycltp
}
\examples{
\dontrun{mycltp(n=20,iter=10000,lambda=10)}
}
