% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{The amount of seats available on the flight}

\item{gamma}{The probability of overbooking}

\item{p}{The probability that a ticketed passenger shows up for the flight}
}
\value{
A list of values for nd, nc, N, p, and gamma, and two plots: discrete and continuous objectives vs n
}
\description{
ntickets
}
\examples{
\dontrun{ntickets(N = 400, gamma = 0.02, p = 0.95)}
}
